requirements.txt
netmiko
PyYAML
netdev